refactor: change streams, technical to tsx
